@@ -9,11 +9,15 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@chakra-ui/react": "^2.1.2",
     "@coinbase/wallet-sdk": "^3.0.5",
+    "@emotion/react": "^11",
+    "@emotion/styled": "^11",
     "@lottiefiles/react-lottie-player": "^3.4.2",
-    "@near-wallet-selector/core": "^3.1.0",
-    "@near-wallet-selector/near-wallet": "^3.1.0",
-    "@near-wallet-selector/sender": "^3.1.0",
+    "@near-wallet-selector/core": "^4.1.0",
+    "@near-wallet-selector/modal-ui": "^4.1.0",
+    "@near-wallet-selector/near-wallet": "^4.1.0",
+    "@near-wallet-selector/sender": "^4.1.0",
     "@nfteyez/sol-rayz": "^0.10.2",
     "@nftpawn-js/core": "../nftpawn-js/dist/packages/core",
     "@reduxjs/toolkit": "^1.8.0",
@@ -33,25 +37,27 @@
     "bootstrap": "^5.1.3",
     "buffer": "^6.0.3",
     "classnames": "^2.3.1",
+    "cleave.js": "^1.6.0",
     "crypto-js": "^4.1.1",
     "ethers": "^5.6.1",
     "final-form": "^4.20.6",
-    "framer-motion": "^6.3.5",
+    "framer-motion": "^6",
     "lodash": "^4.17.21",
     "mobile-detect": "^1.4.5",
     "moment-timezone": "^0.5.34",
     "nanoid": "^3.3.1",
     "near-api-js": "^0.44.2",
     "query-string": "^7.1.1",
-    "rc-switch": "^3.2.2",
-    "react": "^17.0.2",
+    "react": "^18.1.0",
     "react-bootstrap": "^2.1.2",
     "react-content-loader": "^6.1.0",
     "react-copy-to-clipboard": "^5.0.4",
     "react-custom-scrollbars-2": "^4.4.0",
+    "react-datepicker": "^4.8.0",
     "react-device-detect": "^2.1.2",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.1.0",
     "react-final-form": "^6.5.8",
+    "react-icons": "^4.4.0",
     "react-loading-overlay": "^1.0.1",
     "react-redux": "^7.2.6",
     "react-reveal": "^1.2.2",
@@ -65,13 +71,14 @@
     "@rollup/plugin-alias": "^3.1.9",
     "@rollup/plugin-commonjs": "^21.0.2",
     "@rollup/plugin-inject": "^4.0.4",
+    "@types/cleave.js": "^1.4.7",
     "@types/keccak": "^3.0.1",
     "@types/nanoid": "^3.0.0",
     "@types/node": "^17.0.21",
     "@types/react": "^17.0.33",
     "@types/react-bootstrap": "^0.32.29",
     "@types/react-copy-to-clipboard": "^5.0.2",
-    "@types/react-dom": "^17.0.10",
+    "@types/react-dom": "^18.0.5",
     "@types/react-loading-overlay": "^1.0.1",
     "@types/readable-stream": "^2.3.13",
     "@types/sass": "^1.43.1",
