@@ -1,6 +1,5 @@
 import { memo } from 'react';
-import cx from 'classnames';
-import styles from './styles.module.scss';
+import { Flex, Image, Text } from '@chakra-ui/react';
 
 interface CryptoDropdownItemProps {
   icon?: string,
@@ -15,16 +14,14 @@ const CryptoDropdownItem = (props: CryptoDropdownItemProps) => {
   const { name, symbol, balance, network, icon, className } = props;
 
   return (
-    <div className={cx(styles.wrapper, !!balance && styles.haveBalance, className)}>
-      <div>
-        <img alt="" src={icon} className={styles.icon} />
-        {name ? <>{name} <span className={styles.symbol}>{symbol}</span></> : symbol}
-        {/* <span className={styles.network}>{network}</span> */}
-      </div>
-      {!!balance && (
-        <div className={styles.balance}>{balance}</div>
-      )}
-    </div>
+    <Flex fontSize='sm' minW={balance ? 300 : 160} w='100%'  alignItems='center' justifyContent='space-between' className={className}>
+      <Flex alignItems='center'>
+        <Image alt="" src={icon} width='18px' height='18px' mr={4} />
+        <Text mr={2}>{name}</Text>
+        <Text color='text.secondary'>{symbol}</Text>
+      </Flex>
+      {!!balance &&  <Text pr={2}>{balance}</Text>}
+    </Flex>
   );
 };
 
