@@ -1,38 +1,45 @@
-import React, { memo } from "react";
+import React from 'react';
 import { isMobile } from "react-device-detect";
 import { Link } from "react-router-dom";
 import cx from "classnames";
+import { Box, Divider, Flex, Icon, Text, Tooltip } from '@chakra-ui/react';
+import { MdInfoOutline } from 'react-icons/md';
 
 import SocialLinks from 'src/views/apps/socialLinks';
 import { APP_URL } from "src/common/constants/url";
 
 import AppIcon from "../appIcon";
-import MyPopover from '../myPopover';
-import styles from "./styles.module.scss";
+import InfoTooltip from '../infoTooltip';
+import SectionContainer from '../sectionContainer';
 
 const Footer = () => (
-  <div className={cx(isMobile && styles.mobileFooter, styles.wrapper)}>
-    <div className={styles.content}>
-      <div className={styles.left}>
-        <Link to={APP_URL.DISCOVER}>
-          <AppIcon />
-        </Link>
-        <div className={styles.company}>
-          <span className={styles.companyName}><i className="far fa-copyright" />2019-2022 Const LLC.</span>
-          <br />
-          <span>3500 S dupont hwy Dover 19901 Delaware</span>
-        </div>
-      </div>
-      <div className={styles.mid}>First NFTs Lending Solution On Near Protocol</div>
-      <div className={styles.right}>
-        <SocialLinks />
-        <div className={styles.copyright}>
-          <div>This project is in public beta <MyPopover desc="This project is in public beta. - NFTPawn's smart contract is not yet audited by well-known security organization or firm. Use at your own risk!" /></div>
-          <div>Copyright © 2022 NFT Pawn. All rights reserved</div>
-        </div>
-      </div>
-    </div>
-  </div>
+  <Box bgColor='black'>
+    <Divider color='background.darker' />
+    <SectionContainer>
+      <Flex pt={50} pb={30} justifyContent='space-between'>
+        <Box>
+          <Link to={APP_URL.DISCOVER}>
+            <AppIcon />
+          </Link>
+          <Flex mt={4} direction='column' color='text.primary' fontSize='sm'>
+            <Text fontWeight='medium'><i className="far fa-copyright" />2019-2022 Const LLC.</Text>
+            <Text>3500 S dupont hwy Dover 19901 Delaware</Text>
+          </Flex>
+        </Box>
+        <Text fontWeight='medium'>First NFTs Lending Solution On Near Protocol</Text>
+        <Flex direction='column' alignItems='flex-end'>
+          <SocialLinks />
+          <Box mt={4} color='text.secondary' fontSize='small'>
+            <Flex alignItems='center' justifyContent='flex-end'>
+              <Text mr={4}>This project is in public beta</Text>
+              <InfoTooltip label="This project is in public beta. NFT Pawn's smart contract is not yet audited by a well-known security organization or firm. Use at your own risk!" />
+            </Flex>
+            <div>Copyright © 2022 NFT Pawn. All rights reserved</div>
+          </Box>
+        </Flex>
+      </Flex>
+    </SectionContainer>
+  </Box>
 );
 
-export default memo(Footer);
+export default React.memo(Footer);
