@@ -15,6 +15,10 @@ export const APP_URL = {
   TERM_OF_SERVICE: `${BASE_PATH}/terms-of-service`,
   FAQS: `${BASE_PATH}/help`,
   NFT_PAWN_BLOG: `${BASE_PATH}/news/`,
+  DASHBOARD: `${BASE_PATH}/my-assets`,
+  VOTING: `${BASE_PATH}/voting/`,
+  VOTING_PROPOSAL_MAKE: `${BASE_PATH}/voting/proposal/make`,
+  VOTING_DETAIL: `${BASE_PATH}/voting/proposal/detail`,
 }
 
 export const API_URL = { 
@@ -40,15 +44,18 @@ export const API_URL = {
     SYNC_NEAR: `${MCT_ROOT}${MCT_API_BASE}/loans/near/sync`,
     BORROWER_STATS: `${MCT_ROOT}${MCT_API_BASE}/loans/borrower-stats`,
     USER_SETTINGS: `${MCT_ROOT}${MCT_API_BASE}/users/settings`,
+    USER_STATS: `${MCT_ROOT}${MCT_API_BASE}/users/stats`,
     GET_OWNED_NFTS: `${MCT_ROOT}${MCT_API_BASE}/moralis/{owner}/nft`,
+    PWP_BALANCES: `${MCT_ROOT}${MCT_API_BASE}/users/balances/pwp`,
+    PWP_CLAIM: `${MCT_ROOT}${MCT_API_BASE}/users/balances/claim`,
+    BALANCES_TRANSACTIONS: `${MCT_ROOT}${MCT_API_BASE}/users/balances/transactions`,
+
+    VOTING_CURRENCIES: `${MCT_ROOT}${MCT_API_BASE}/currencies/pwp-token`,
+    VOTING_PROPOSAL: `${MCT_ROOT}${MCT_API_BASE}/proposals/create`,
+    VOTING_PROPOSAL_LIST: `${MCT_ROOT}${MCT_API_BASE}/proposals/list`,
+    VOTING_PROPOSAL_DETAIL: `${MCT_ROOT}${MCT_API_BASE}/proposals/detail`,
+    VOTING_PROPOSAL_DETAIL_VOTES: `${MCT_ROOT}${MCT_API_BASE}/proposals/votes/list`,
+    VOTING_PROPOSAL_VOTE_CREATE: `${MCT_ROOT}${MCT_API_BASE}/proposals/votes/create`,
+    VOTING_PROPOSAL_VOTE_CHECK: `${MCT_ROOT}${MCT_API_BASE}/proposals/votes/vote`,
   },
 }
-
-// export const ALCHEMY_URL = {
-//   GET_NFTS: APP_CLUSTER === 'mainnet' ? '' : `https://polygon-mumbai.g.alchemy.com/v2/${ALCHEMY_API_KEY}/getNFTs/`,
-//   GET_NFT_METADATA: APP_CLUSTER === 'mainnet' ? '' : `https://polygon-mumbai.g.alchemy.com/v2/${ALCHEMY_API_KEY}/getNFTMetadata/`,
-// }
-
-// export const MORALIS_URL = {
-//   GET_NFTS: `https://deep-index.moralis.io/api/v2/{owner}/nft`,
-// }
