@@ -55,7 +55,7 @@ export default class PayLoanNearTransaction extends NearTransaction {
 
       const res = await window.nearSelector.signAndSendTransactions({ 
         transactions,
-        callbackUrl: this.generateCallbackUrl({ token_id: assetTokenId, contract_address: assetContractAddress }, `${window.location.origin}${APP_URL.MY_NFT}`),
+        callbackUrl: this.generateCallbackUrl({ token_id: assetTokenId, contract_address: assetContractAddress }, `${window.location.origin}${APP_URL.DASHBOARD}/loans`),
       });
       
       return this.handleSuccess(
