@@ -2,20 +2,21 @@ import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import cx from 'classnames';
 import NftPawn, { CollectionData } from '@nftpawn-js/core';
+import { isMobile } from 'react-device-detect';
+import { Box, Heading, Text } from '@chakra-ui/react';
 
 import { APP_URL } from 'src/common/constants/url';
 import BodyContainer from 'src/common/components/bodyContainer';
+import { CollectionNft } from 'src/modules/nftLend/models/collection';
 
 import Item from './item';
 import styles from './styles.module.scss';
-import { isMobile } from 'react-device-detect';
-import { CollectionNft } from 'src/modules/nftLend/models/collection';
 
 export const OnBoardingHeader = () => (
-  <div className={cx(isMobile && styles.mbHeader, styles.headerWrapper)}>
-    <h5>Create, explore, & collect digital art NFTs</h5>
-    <h1>The new creative economy</h1>
-  </div>
+  <Box textAlign='center' p={20}>
+    <Text fontSize='xs' color='text.secondary' fontWeight='bold' letterSpacing='wider'>UNLOCK THE NEW UTILITY FOR NFTS</Text>
+    <Heading as='h1' letterSpacing='wider' textTransform='uppercase'>The new creative economy</Heading>
+  </Box>
 );
 
 const Discover = () => {
