@@ -1,5 +1,7 @@
 import BigNumber from "bignumber.js";
 import React from "react";
+import { MdInfoOutline } from 'react-icons/md';
+import { Flex, Grid, GridItem, Icon, Text, Tooltip } from '@chakra-ui/react';
 
 import icPriceTag from "../images/ic_price_tag.svg";
 import LoanDetailButtons from "./LoanDetail.Buttons";
@@ -8,11 +10,10 @@ import { LoanNft } from 'src/modules/nftLend/models/loan';
 import { isSameAddress } from 'src/common/utils/helper';
 import { useCurrentWallet } from 'src/modules/nftLend/hooks/useCurrentWallet';
 import { LOAN_DURATION } from 'src/modules/nftLend/constant';
-import { formatCurrencyByLocale } from "src/common/utils/format";
+import { formatCurrency } from "src/common/utils/format";
 
 import pawnStyles from './pawnInfo.module.scss';
 import styles from "../styles.module.scss";
-import CountdownText from 'src/common/components/countdownText';
 
 export interface LoanDetailProps {
   loan: LoanNft;
@@ -35,19 +36,26 @@ const LoanDetailPriceInfo: React.FC<LoanDetailPriceInfoProps> = ({ loan }) => {
     <div className={styles.infoPrice}>
       <div className={pawnStyles.head}> 
         <div>
-          <div className={styles.infoPriceTags}>
+          <Flex className={styles.infoPriceTags}>
             <label>Principal</label>
             <img src={icPriceTag} alt="item price" />
-          </div>
+          </Flex>
           <div className={styles.infoPriceValue}>
-            <div>{`${formatCurrencyByLocale(
+            <div>{`${formatCurrency(
               loan.principal_amount,
               2
             )} ${loan.currency?.symbol}`}</div>
           </div>
         </div>
         <div className={pawnStyles.configs}>
-          <label>Negotiation</label>
+          <Flex gap={2} alignItems='center' color='text.secondary' >
+            <Text color='text.secondary' fontWeight='semibold' fontSize='sm'>
+              Negotiation Terms
+            </Text>
+            <Tooltip placement='top' label='The following terms are negotiable: Principal, duration, and interest rate. You can set your own terms here.'>
+              <span><Icon as={MdInfoOutline} /></span>
+            </Tooltip>
+          </Flex>
           <ul> 
             <li className={loan.isAllowChange('principal_amount') ? pawnStyles.allow : pawnStyles.notallow }>Principal</li>
             <li className={loan.isAllowChange('duration') ? pawnStyles.allow : pawnStyles.notallow }>Duration</li>
@@ -56,72 +64,51 @@ const LoanDetailPriceInfo: React.FC<LoanDetailPriceInfoProps> = ({ loan }) => {
         </div>
       </div>
       <LoanDetailButtons loan={loan} userOffer={userOffer} />
-      <div className={styles.feeInfoWrap}>
-        <div>
-          <div className={styles.feeInfoTitle}>Interest rate</div>
-          <div className={styles.feeInfoValue}>
-            {new BigNumber(loan.interest_rate)
-              .multipliedBy(100)
-              .toNumber()}
-            %
-          </div>
-        </div>
-        <div>
-          <div className={styles.feeInfoTitle}>Terms</div>
-          <div className={styles.feeInfoValue}>
+      <Grid templateColumns='repeat(5, 1fr)'>
+        <GridItem>
+          <Text fontSize='sm' color='text.secondary'>Interest rate</Text>
+          <Text fontWeight='semibold'>
+            {new BigNumber(loan.interest_rate).multipliedBy(100).toNumber()}%
+          </Text>
+        </GridItem>
+        <GridItem>
+          <Text fontSize='sm' color='text.secondary'>Terms</Text>
+          <Text fontWeight='semibold'>
             {loanDuration ? loanDuration.label : `${Math.ceil(new BigNumber(loan.duration).dividedBy(86400).toNumber())} days`}
-          </div>
-        </div>
+          </Text>
+        </GridItem>
         {userOffer && (
           <>
-            <div>
-              <div
-                className={styles.feeInfoTitle}
-                style={{ color: "green", opacity: 0.8 }}
-              >
+            <GridItem>
+              <Text fontSize='sm' color='brand.success.700'>
                 My Principal
-              </div>
-              <div
-                className={styles.feeInfoValue}
-                style={{ color: "green", fontWeight: "bold" }}
-              >
+              </Text>
+              <Text fontWeight='semibold' color='brand.success.700'>
                 {userOffer?.principal_amount} {loan.currency?.symbol}
-              </div>
-            </div>
-            <div>
-              <div
-                className={styles.feeInfoTitle}
-                style={{ color: "green", opacity: 0.8 }}
-              >
+              </Text>
+            </GridItem>
+            <GridItem>
+              <Text fontSize='sm' color='brand.success.700'>
                 My Interest rate
-              </div>
-              <div
-                className={styles.feeInfoValue}
-                style={{ color: "green", fontWeight: "bold" }}
-              >
+              </Text>
+              <Text fontWeight='semibold' color='brand.success.700'>
                 {new BigNumber(userOffer?.interest_rate)
                   .multipliedBy(100)
                   .toNumber()}
                 %
-              </div>
-            </div>
-            <div>
-              <div
-                className={styles.feeInfoTitle}
-                style={{ color: "green", opacity: 0.8 }}
-              >
+              </Text>
+            </GridItem>
+            <GridItem>
+              <Text fontSize='sm' color='brand.success.700'>
                 My Terms
-              </div>
-              <div
-                className={styles.feeInfoValue}
-                style={{ color: "green", fontWeight: "bold" }}
-              >
+              </Text>
+              <Text fontWeight='semibold' color='brand.success.700'>
                 {offerDuration ? offerDuration.label : `${Math.ceil(new BigNumber(userOffer.duration).dividedBy(86400).toNumber())} days`}
-              </div>
-            </div>
+              </Text>
+            </GridItem>
           </>
         )}
-      </div>
+      </Grid>
     </div>
   );
 };
