@@ -1,37 +1,49 @@
 import React from "react";
-import cx from "classnames";
 import BigNumber from 'bignumber.js';
+import { Grid, GridItem } from '@chakra-ui/react';
 
 import { LoanNft } from 'src/modules/nftLend/models/loan';
 import { formatCurrency } from 'src/common/utils/format';
-
-import styles from "../styles.module.scss";
+import BoxAttrValue from 'src/views/loanDetail/BoxAttrValue';
 
 interface LoanDetailPriceStatisticProps {
   loan: LoanNft;
 };
 
 const LoanDetailPriceStatistic: React.FC<LoanDetailPriceStatisticProps> = ({ loan }) => {
-  const usdValue = new BigNumber(loan.asset?.stats?.avg_price).multipliedBy(loan.asset?.stats?.currency?.price);
-  const ltv = !usdValue.isEqualTo(0)
-    ? new BigNumber(loan.principal_amount).multipliedBy(loan.currency?.price).dividedBy(usdValue).multipliedBy(100).toNumber()
-    : 0;
+  const principalUsdValue = new BigNumber(loan.principal_amount).multipliedBy(loan.currency?.price);
+  const avgUsdValue = new BigNumber(loan.asset?.stats?.avg_price).multipliedBy(loan.asset?.stats?.currency?.price);
+  const floorUsdValue = new BigNumber(loan.asset?.stats?.floor_price).multipliedBy(loan.asset?.stats?.currency?.price);
+  const ltv = avgUsdValue.isGreaterThan(0)
+    ? principalUsdValue.dividedBy(avgUsdValue).multipliedBy(100).toNumber()
+    : floorUsdValue.isGreaterThan(0)
+      ? principalUsdValue.dividedBy(floorUsdValue).multipliedBy(100).toNumber()
+      : 0
 
   return (
-    <div className={cx(styles.tabContentWrap, styles.tabContentAttrWrap)}>
-      <div className={styles.tabContentAttrItem}>
-        <label>Avg Price</label>
-        <div>{formatCurrency(loan.asset?.stats?.avg_price)} {loan.asset?.stats?.currency?.symbol}</div>
-      </div>
-      <div className={styles.tabContentAttrItem}>
-        <label>Floor Price</label>
-        <div>{formatCurrency(loan.asset?.stats?.floor_price)} {loan.asset?.stats?.currency?.symbol}</div>
-      </div>
-      <div className={styles.tabContentAttrItem}>
-        <label>Loan To Value</label>
-        <div>{ltv ? `${formatCurrency(ltv)}%` : 'Not Available'}</div>
-      </div>
-    </div>
+    <Grid templateColumns={{ md: 'repeat(3, 1fr)' }} gap={2}>
+      <GridItem>
+        <BoxAttrValue
+          label='Avg Price'
+          value={`${formatCurrency(loan.asset?.stats?.avg_price)} ${loan.asset?.stats?.currency?.symbol}`}
+          desc='The average price is a representative measure of a range of prices. It is calculated by taking the sum of NFT values and dividing it by the number of NFT listied.'
+        />
+      </GridItem>
+      <GridItem>
+        <BoxAttrValue
+          label='Floor Price'
+          value={`${formatCurrency(loan.asset?.stats?.floor_price)} ${loan.asset?.stats?.currency?.symbol}`}
+          desc={`The floor price is the lowest 'Buy Now' price for an NFT within a collection`}
+        />
+      </GridItem>
+      <GridItem>
+        <BoxAttrValue
+          label='Loan to value'
+          value={ltv ? `${formatCurrency(ltv)}%` : 'Not Available'}
+          desc={`Lenders use the loan-to-value (LTV) ratio to determine how much risk they're taking on with a secured loan. Any NFTPawn valuation metric is based on estimates and metrics that do not represent financial advice or the actual expected LTV of an NFT listing. Please conduct your own research.`}
+        />
+      </GridItem>
+    </Grid>
   );
 };
 
