@@ -1,265 +1,148 @@
-import React from "react";
-import cx from "classnames";
+import { Flex, Heading, ListItem, OrderedList, Text, UnorderedList } from '@chakra-ui/react'
 
-import BodyContainer from "src/common/components/bodyContainer";
-
-import styles from "./styles.module.scss";
-import { isMobile } from "react-device-detect";
+import BodyContainer from "src/common/components/bodyContainer"
+import styles from './styles.module.scss'
 
 const TermsOfService = () => {
   return (
-    <BodyContainer
-      className={cx(
-        isMobile && styles.mbWrapper,
-        styles.wrapper,
-        styles.wrapper
-      )}
-    >
-      <div>
-        <h3>NFT Pawn Platform Agreement</h3>
-        <p>(Last updated FEBRUARY 22, 2022)</p>
-        <p>
-          Please read this NFT Pawn Platform Agreement (the "Agreement")
-          carefully. This Agreement is a legally binding agreement between you
-          (" you", "your" or "User") and Non-fungible Technologies, Inc.
-          ("MyConstant" or "we" or "us"). By clicking or tapping any button or
-          box marked "I Accept," "I Agree," or "OK" (or a similar term) in
-          connection with this Agreement, or by using or accessing the NFT Pawn
-          Platform (as defined below), you acknowledge and agree to this
-          Agreement and you confirm that you are of legal age where you live and
-          have the legal capacity to enter into this Agreement.
-        </p>
-        <p>
-          This Agreement is between you and MyConstant concerning your use of
-          certain MyConstant services in connection with its decentralized,
-          peer-to-peer lending platform that allows the creation of loans
-          collateralized by non-fungible tokens ("NFTs"), fungible tokens and/or
-          other crypto-assets (the " NFT Pawn Platform"). YOU ACKNOWLEDGE AND
-          AGREE THAT NEITHER MyConstant NOR THE PLATFORM IS A LENDER OR BORROWER
-          OF LOANS, AND THAT YOUR USE OF THE PLATFORM DOES NOT MAKE MyConstant
-          OR THE PLATFORM A LENDER OR BORROWER OF LOANS.
-        </p>
-        <p>
-          If you are an individual accessing or using the NFT Pawn Platform on
-          behalf of, or for the benefit of, any corporation, partnership or
-          other entity with which you are associated (an "Organization"), then
-          you are agreeing to this Agreement on behalf of yourself and such
-          Organization, and you represent and warrant that you have the legal
-          authority to bind such Organization to this Agreement. References to
-          "you" and "User" in this Agreement will refer to both the individual
-          using the NFT Pawn Platform and to any such Organization.
-        </p>
-        <p>
-          THIS AGREEMENT CONTAINS A MANDATORY ARBITRATION PROVISION THAT, AS
-          FURTHER SET FORTH IN SECTION 19 BELOW, REQUIRES THE USE OF ARBITRATION
-          ON AN INDIVIDUAL BASIS TO RESOLVE DISPUTES, RATHER THAN JURY TRIALS OR
-          ANY OTHER COURT PROCEEDINGS, OR CLASS ACTIONS OF ANY KIND.
-        </p>
-        <ol>
-          <li>
-            <b>Changes</b>
-            <ul>
-              <li>
-                We may change, modify or supplement this Agreement from time to
-                time by notifying you of such changes by any reasonable means,
-                including by posting a modified Agreement through the NFT Pawn
-                Platform. No such changes will apply to any dispute between you
-                and us arising before the date on which we posted the revised
-                Agreement incorporating such changes, or otherwise notified you
-                of such changes. The "Last Updated" legend above indicates when
-                this Agreement was last changed. Your use of the NFT Pawn
-                Platform after any changes to this Agreement will constitute
-                your acceptance of such changes. We may, at any time and without
-                liability: (a) modify or discontinue all or part of the NFT-Lend
-                Platform; (b) charge, modify or waive any fees required to use
-                the NFT Pawn Platform; or (c) offer opportunities to some or
-                all users.
-              </li>
-              <li>
-                You may not use the NFT Pawn Platform outside the scope set
-                forth in this Agreement without the prior written consent of
-                MyConstant, which may be withheld in MyConstant's sole
-                discretion. MyConstant reserves the right to refuse, modify, or
-                terminate access to or use of all or part of the NFT Pawn
-                Platform to anyone for any reason at our discretion.
-              </li>
-            </ul>
-          </li>
-          <li>
-            <b>Registration; Accounts.</b>
-            <ul>
-              <li>
-                You may need to register to use all or part of the NFT Pawn
-                Platform. You must accurately and completely supply all
-                requested information in order to complete the registration
-                process. You represent and warrant that any information you
-                provide in connection with the NFT Pawn Platform is and will
-                remain accurate and complete at all times during your use of and
-                access to the NFT Pawn Platform, and that you will maintain and
-                update such information as needed. We may also establish
-                additional or different requirements from time to time, by
-                posting such requirements on our website or otherwise
-                communicating them to you.
-              </li>
-              <li>
-                When your account becomes active on the NFT Pawn Platform, you
-                may have rights and responsibilities as a potential requestor of
-                a loan (a " Borrower") or a potential provider of such loans (a
-                "Lender"), or both. We may reject, or require that you change,
-                any username, password or other information that you provide to
-                us in registering. Your username and password are for your
-                personal use only and should be kept confidential. You, and not
-                MyConstant, are responsible for any use or misuse of your
-                username or password, and you must promptly notify us of any
-                breach or unauthorized use of or access to your username or
-                password, or your account. You will be solely liable for all
-                activity arising from your access to, use of, or any other act
-                or omission with respect to the NFT Pawn Platform, whether or
-                not authorized by you (including any acts or omissions by any
-                users for whom you have submitted a user authorization form).
-              </li>
-              <li>
-                You are responsible for obtaining and maintaining all hardware,
-                software (and related licenses) and communications equipment
-                necessary to access and use the NFT Pawn Platform and for
-                paying all third-party access charges (e.g., ISP,
-                telecommunications, etc.) incurred while using the NFT Pawn
-                Platform. You acknowledge that you have been advised of and can
-                comply with all minimum networking, hardware, software (and
-                related licenses), firewalls and/or environmental conditions,
-                and communications requirements applicable to the NFT Pawn
-                Platform.
-              </li>
-            </ul>
-          </li>
-          <li>
-            <b>Rules of Conduct.</b>
-            <ul>
-              <li>
+    <BodyContainer>
+      <Flex className={styles.wrapper} gap={4} direction='column' color='text.primary' fontSize='sm' lineHeight='6'>
+        <Heading as='h1'>NFT Pawn Platform Agreement</Heading>
+        <Text fontSize='sm' fontWeight='semibold' color='text.secondary'>(Last updated FEBRUARY 22, 2022)</Text>
+        <Text fontWeight='medium'>
+          Please read this NFT Pawn Platform Agreement (the "Agreement") carefully. This Agreement is a legally binding agreement between you (" you", "your" or "User") and Non-fungible Technologies, Inc. ("NFT Pawn" or "we" or "us"). By clicking or tapping any button or box marked "I Accept," "I Agree," or "OK" (or a similar term) in connection with this Agreement, or by using or accessing the NFT Pawn Platform (as defined below), you acknowledge and agree to this Agreement and you confirm that you are of legal age where you live and have the legal capacity to enter into this Agreement.
+        </Text>
+        <Text fontWeight='bold'>
+          This Agreement is between you and NFT Pawn concerning youruse of certain NFT Pawn services in connection with its decentralized, peer-to-peer lending platform that allows the creation of loans collateralized by non-fungible tokens ("NFTs"), fungible tokens and/or other crypto-assets (the " NFT Pawn Platform"). YOU ACKNOWLEDGE AND AGREE THAT NEITHER NFT Pawn NOR THE PLATFORM IS A LENDER OR BORROWER OF LOANS, AND THAT YOUR USE OF THE PLATFORM DOES NOT MAKE NFT Pawn OR THE PLATFORM A LENDER OR BORROWER OF LOANS.
+        </Text>
+        <Text fontWeight='medium'>
+          If you are an individual accessing or using the NFT Pawn Platform on behalf of, or for the benefit of, any corporation, partnership or other entity with which you are associated (an "Organization"), then you are agreeing to this Agreement on behalf of yourself and such Organization, and you represent and warrant that you have the legal authority to bind such Organization to this Agreement. References to "you" and "User" in this Agreement will refer to both the individual using the NFT Pawn Platform and to any such Organization. 
+        </Text>
+        <Text fontWeight='medium'>
+          THIS AGREEMENT CONTAINS A MANDATORY ARBITRATION PROVISION THAT, AS FURTHER SET FORTH IN SECTION 19 BELOW, REQUIRES THE USE OF ARBITRATION ON AN INDIVIDUAL BASIS TO RESOLVE DISPUTES, RATHER THAN JURY TRIALS OR ANY OTHER COURT PROCEEDINGS, OR CLASS ACTIONS OF ANY KIND.
+        </Text>
+        <OrderedList>
+          <ListItem fontSize='lg'>
+            <Heading as='h2' fontSize='lg' mb={2}>Changes</Heading>
+            <UnorderedList fontSize='sm'>
+              <ListItem>
+                We may change, modify or supplement this Agreement from time to time by notifying you of such changes by any reasonable means, including by posting a modified Agreement through the NFT Pawn Platform. No such changes will apply to any dispute between you and us arising before the date on which we posted the revised Agreement incorporating such changes, or otherwise notified you of such changes. The "Last Updated" legend above indicates when this Agreement was last changed. Your use of the NFT Pawn Platform after any changes to this Agreement will constitute your acceptance of such changes. We may, at any time and without liability: (a) modify or discontinue all or part of the NFT Pawn Platform; (b) charge, modify or waive any fees required to use the NFT Pawn Platform; or (c) offer opportunities to some or all users.
+              </ListItem>
+              <ListItem>
+                You may not use the NFT Pawn Platform outside the scope set forth in this Agreement without the prior written consent of NFT Pawn, which may be withheld in NFT Pawn's sole discretion. NFT Pawn reserves the right to refuse, modify, or terminate access to or use of all or part of the NFT Pawn Platform to anyone for any reason at our discretion.
+              </ListItem>
+            </UnorderedList>
+          </ListItem>
+          <ListItem fontSize='lg'>
+            <Heading as='h2' fontSize='lg' mb={2}>Registration; Accounts.</Heading>
+            <UnorderedList fontSize='sm'>
+              <ListItem>
+                You may need to register to use all or part of the NFT Pawn Platform. You must accurately and completely supply all requested information in order to complete the registration process. You represent and warrant that any information you provide in connection with the NFT Pawn Platform is and will remain accurate and complete at all times during your use of and access to the NFT Pawn Platform, and that you will maintain and update such information as needed. We may also establish additional or different requirements from time to time, by posting such requirements on our website or otherwise communicating them to you.
+              </ListItem>
+              <ListItem>
+                When your account becomes active on the NFT Pawn Platform, you may have rights and responsibilities as a potential requestor of a loan (a " Borrower") or a potential provider of such loans (a "Lender"), or both. We may reject, or require that you change, any username, password or other information that you provide to us in registering. Your username and password are for your personal use only and should be kept confidential. You, and not NFT Pawn, are responsible for any use or misuse of your username or password, and you must promptly notify us of any breach or unauthorized use of or access to your username or password, or your account. You will be solely liable for all activity arising from your access to, use of, or any other act or omission with respect to the NFT Pawn Platform, whether or not authorized by you (including any acts or omissions by any users for whom you have submitted a user authorization form). 
+              </ListItem>
+              <ListItem>
+                You are responsible for obtaining and maintaining all hardware, software (and related licenses) and communications equipment necessary to access and use the NFT Pawn Platform and for paying all third-party access charges (e.g., ISP, telecommunications, etc.) incurred while using the NFT Pawn Platform. You acknowledge that you have been advised of and can comply with all minimum networking, hardware, software (and related licenses), firewalls and/or environmental conditions, and communications requirements applicable to the NFT Pawn Platform.
+              </ListItem>
+            </UnorderedList>
+          </ListItem>
+          <ListItem fontSize='lg'>
+            <Heading as='h2' fontSize='lg' mb={2}>Rules of Conduct.</Heading>
+            <UnorderedList fontSize='sm'>
+              <ListItem>
                 In connection with the NFT Pawn Platform, you must not:
-                <ul>
-                  <li>
-                    Borrow or lend in a manner for which you are not authorized
-                    or licensed to do so, or for which you are not permitted to
-                    do so either by authorization or licensure or by any
-                    applicable federal, state, local or foreign laws (including
-                    common law), codes, statutes, ordinances, rules,
-                    regulations, regulatory bulletins, interpretations or
-                    guidance or orders, decrees and orders of, official releases
-                    or any other requirements by, any governmental authority or
-                    agency (collectively, "Applicable Law");
-                  </li>
-                  <li>
-                    Use the NFT Pawn Platform for any purpose or in connection
-                    with any activity that is fraudulent, illegal or otherwise
-                    tortious or unlawful;
-                  </li>
-                  <li>
-                    In connection with your access and use of the NFT Pawn
-                    Platform, discriminate against any Borrower member or group
-                    on the basis of race, color, religion, national origin, sex,
-                    marital status, age, sexual orientation, military status,
-                    the Borrower member's source of income, or any other basis
-                    prohibited by any Applicable Law, including the Equal Credit
-                    Opportunity Act and its implementing regulation, Regulation
-                    B or the Fair Housing Act;
-                  </li>
-                  <li>
-                    Post, transmit or otherwise make available through or in
-                    connection with the NFT Pawn Platform any materials that
-                    are or may be: (a) threatening, harassing, degrading,
-                    hateful or intimidating, or otherwise fail to respect the
-                    rights and dignity of others; (b) defamatory, libelous,
-                    fraudulent or otherwise tortious or illegal, including in
-                    connection with the disclosure of insider information under
-                    securities law or of another party's trade secrets; (c)
-                    obscene, indecent, pornographic or otherwise objectionable;
-                    (d) protected by copyright, trademark, trade secret, right
-                    of publicity or privacy or any other proprietary right,
-                    without the express prior written consent of the applicable
-                    owner; (e) information or content that you do not have a
-                    right to make available under any law or under contractual
-                    or fiduciary relationships; or (f) information or content
-                    that you know is not correct and current;
-                  </li>
-                  <li>
+                <UnorderedList>
+                  <ListItem>
+                    Borrow or lend in a manner for which you are not authorized or licensed to do so, or for which you are not permitted to do so either by authorization or licensure or by any applicable federal, state, local or foreign laws (including common law), codes, statutes, ordinances, rules, regulations, regulatory bulletins, interpretations or guidance or orders, decrees and orders of, official releases or any other requirements by, any governmental authority or agency (collectively, "Applicable Law");
+                  </ListItem>
+                  <ListItem>
+                    Use the NFT Pawn Platform for any purpose or in connection with any activity that is fraudulent, illegal or otherwise tortious or unlawful;
+                  </ListItem>
+                  <ListItem>
+                    In connection with your access and use of the NFT Pawn Platform, discriminate against any Borrower member or group on the basis of race, color, religion, national origin, sex, marital status, age, sexual orientation, military status, the Borrower member's source of income, or any other basis prohibited by any Applicable Law, including the Equal Credit Opportunity Act and its implementing regulation, Regulation B or the Fair Housing Act;
+                  </ListItem>
+                  <ListItem>
+                    Post, transmit or otherwise make available through or in connection with the NFT Pawn Platform any materials that are or may be: (a) threatening, harassing, degrading, hateful or intimidating, or otherwise fail to respect the rights and dignity of others; (b) defamatory, libelous, fraudulent or otherwise tortious or illegal, including in connection with the disclosure of insider information under securities law or of another party's trade secrets; (c) obscene, indecent, pornographic or otherwise objectionable; (d) protected by copyright, trademark, trade secret, right of publicity or privacy or any other proprietary right, without the express prior written consent of the applicable owner; (e) information or content that you do not have a right to make available under any law or under contractual or fiduciary relationships; or (f) information or content that you know is not correct and current;
+                  </ListItem>
+                  <ListItem>
                     Send unlawful unsolicited commercial messages through the
                     NFT Pawn Platform;
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     Post, transmit or otherwise make available any virus, worm,
                     Trojan horse, Easter egg, time bomb, spyware or other
                     computer code, file or program that is or is potentially
                     harmful or invasive or intended to damage or hijack the
                     operation of, or to monitor the use of, any hardware,
                     software or equipment (each, a "Virus");
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     Interfere with or disrupt the operation of the NFT Pawn
                     Platform or the servers or networks used to make the
                     NFT Pawn Platform available, including by hacking or
                     defacing any portion of the NFT Pawn Platform; or violate
                     any requirement, procedure or policy of such servers or
                     networks;
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     Perform testing of the NFT Pawn Platform, or the systems or
                     networks used to provide the NFT Pawn Platform, to attempt
                     to find vulnerabilities;
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     Run a computer program, a set of computer programs, or other
                     operations to attempt to assess the relative performance of
                     the NFT Pawn Platform, or the systems or networks used to
                     provide the NFT Pawn Platform;
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     Restrict or inhibit any other person from using the
                     NFT Pawn Platform;
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     Harvest or collect information about users of the NFT Pawn
                     Platform;
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     Remove any copyright, trademark or other proprietary rights
-                    notice from the NFT Pawn Platform or any MyConstant
+                    notice from the NFT Pawn Platform or any NFT Pawn
                     Material (defined in Section 8 below);
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     Reproduce, modify, adapt, translate, create derivative works
                     of, sell, resell, rent, lease, loan, timeshare, distribute
                     or otherwise exploit any portion of (or any use of) the
-                    NFT Pawn Platform, or any MyConstant Material, except as
+                    NFT Pawn Platform, or any NFT Pawn Material, except as
                     expressly authorized herein;
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     Reverse engineer, decompile or disassemble, or attempt to
                     discover the source code of underlying ideas or algorithms
-                    of, any portion of the NFT Pawn Platform, or any MyConstant
+                    of, any portion of the NFT Pawn Platform, or any NFT Pawn
                     Material, except where such restriction is expressly
                     prohibited by Applicable Law;
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     Frame or mirror any portion of the NFT Pawn Platform, or
                     otherwise incorporate any portion of the NFT Pawn Platform
-                    or MyConstant Material into any product or service;
-                  </li>
-                  <li>
-                    Systematically download and store MyConstant Material;
-                  </li>
-                  <li>
+                    or NFT Pawn Material into any product or service;
+                  </ListItem>
+                  <ListItem>
+                    Systematically download and store NFT Pawn Material;
+                  </ListItem>
+                  <ListItem>
                     Use any robot, spider, site search/retrieval application or
                     other manual or automatic device to retrieve, index,
-                    "scrape," "data mine" or otherwise gather MyConstant
+                    "scrape," "data mine" or otherwise gather NFT Pawn
                     Material, or reproduce or circumvent the navigational
                     structure or presentation of the NFT Pawn Platform;
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     Access or use the NFT Pawn Platform in a way intended to
                     avoid incurring fees or exceeding usage limits or quotas;
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     Conduct or attempt to conduct a transaction or loan outside
                     of the NFT Pawn Platform that was initiated through the
                     NFT Pawn Platform (or conduct or attempt to conduct a loan
@@ -268,32 +151,32 @@ const TermsOfService = () => {
                     attempt to conduct business relating to loans outside of the
                     NFT Pawn Platform, with a Lender or Borrower that you met
                     through the NFT Pawn Platform; or
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     Cause, assist or otherwise encourage any third party to do
                     any of the foregoing.
-                  </li>
-                </ul>
-                <li>
+                  </ListItem>
+                </UnorderedList>
+                <ListItem>
                   We reserve the right in our sole discretion to terminate or
                   suspend any account that violates the foregoing or any other
                   provision of this Agreement, and any account that has been
                   inactive for a substantial period of time.
-                </li>
-              </li>
-            </ul>
-          </li>
-          <li>
-            <b>Listings; Transactions.</b>
-            <ul>
-              <li>
+                </ListItem>
+              </ListItem>
+            </UnorderedList>
+          </ListItem>
+          <ListItem>
+            <Heading as='h2' fontSize='lg' mb={2}>Listings; Transactions.</Heading>
+            <UnorderedList>
+              <ListItem>
                 The NFT Pawn Platform may make available the opportunity for
                 Borrowers to submit requests, proposals, terms, descriptions and
                 other information in connection with potential loans
                 ("Listing"). It is your responsibility to ascertain and obey all
                 Applicable Laws regarding Listings.
-              </li>
-              <li>
+              </ListItem>
+              <ListItem>
                 The NFT Pawn Platform may also make available the opportunity
                 for Lenders to accept a Borrower's loan request contained in the
                 Listing and transfer funds to the Borrower ("Transaction"). If
@@ -303,16 +186,16 @@ const TermsOfService = () => {
                 information to third parties for purposes of facilitating
                 Transactions. Verification of information may be required prior
                 to the acknowledgment or completion of any Transaction.
-              </li>
-              <li>
+              </ListItem>
+              <ListItem>
                 LISTINGS ARE MADE AVAILABLE BY YOU AND BY THIRD PARTIES, AND NOT
-                BY MyConstant. WE MAKE NO REPRESENTATIONS AS TO THE
+                BY NFT Pawn. WE MAKE NO REPRESENTATIONS AS TO THE
                 COMPLETENESS, ACCURACY, RELIABILITY, VALIDITY OR TIMELINESS OF
                 LISTINGS OR THE ENFORCEABILITY OF THE LOANS. TRANSACTIONS ARE
-                SOLELY BETWEEN YOU AND OTHER USERS. MyConstant IS NOT A LENDER
+                SOLELY BETWEEN YOU AND OTHER USERS. NFT Pawn IS NOT A LENDER
                 OR BORROWER OR OTHERWISE A PARTY TO ANY TRANSACTION, AND IN
                 CONNECTION WITH YOUR REQUEST FOR OR PROVISION OF ANY LOANS,
-                MyConstant IS NOT ACTING IN ANY AGENCY OR FIDUCIARY CAPACITY AND
+                NFT Pawn IS NOT ACTING IN ANY AGENCY OR FIDUCIARY CAPACITY AND
                 ASSUMES NO ADVISORY OR FIDUCIARY RESPONSIBILITY THEREWITH. YOU
                 ARE SOLELY RESPONSIBLE FOR YOUR INTERACTIONS WITH OTHER USERS,
                 AND FOR ANY TRANSACTIONS. WE RESERVE THE RIGHT, BUT HAVE NO
@@ -324,154 +207,154 @@ const TermsOfService = () => {
                 TRANSACTIONS; (D) YOUR OR ANY OTHER USER'S ACTION, INACTION OR
                 BREACH OF OUR TERMS; OR (E) OUR ENFORCEMENT OF (OR FAILURE TO
                 ENFORCE) OUR TERMS AGAINST YOU OR ANY OTHER USER.
-              </li>
-              <li>
-                For the avoidance of doubt, MyConstant will have no obligation
+              </ListItem>
+              <ListItem>
+                For the avoidance of doubt, NFT Pawn will have no obligation
                 to provide any consulting, implementation, or other professional
                 services, except as may be set forth in a separate written
-                agreement between you and MyConstant.
-              </li>
-            </ul>
-          </li>
-          <li>
-            <b>Borrower Obligations.</b>
-            <ul>
-              <li>
+                agreement between you and NFT Pawn.
+              </ListItem>
+            </UnorderedList>
+          </ListItem>
+          <ListItem>
+            <Heading as='h2' fontSize='lg' mb={2}>Borrower Obligations.</Heading>
+            <UnorderedList>
+              <ListItem>
                 In addition to your other obligations under this Agreement, this
                 Section 5 is applicable to you in your capacity as a Borrower.
-              </li>
-              <li>
-                (a) You agree to comply with MyConstant's Borrower policies and
+              </ListItem>
+              <ListItem>
+                (a) You agree to comply with NFT Pawn's Borrower policies and
                 practices, as set forth herein, as may be amended from time to
                 time, and agree that:
-                <ul>
-                  <li>
+                <UnorderedList>
+                  <ListItem>
                     You assume full responsibility for the loan proposal(s)
                     offered and the accuracy and detail of the Listing(s), and
                     you warrant that all information contained in each of your
                     Listings and related loans is correct, accurate, verifiable
                     and not misleading;
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     You are prohibited from attempting to complete or completing
                     outside of the NFT Pawn Platform any Listing or Transaction
                     listed in connection with the NFT Pawn Platform;
-                  </li>
-                  <li>
-                    MyConstant may revise data associated with Listings to
+                  </ListItem>
+                  <ListItem>
+                    NFT Pawn may revise data associated with Listings to
                     supplement, remove, or correct information as deemed
                     necessary, while maintaining the integrity of the Listing,
-                    and you shall promptly notify MyConstant in writing if you
+                    and you shall promptly notify NFT Pawn in writing if you
                     believe that any such supplementation, removal or correction
                     is inaccurate, incomplete or misleading;
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     If a loan proposal is accepted by a Lender, you will proceed
                     according to the loan terms;
-                  </li>
-                  <li>
-                    You, and not MyConstant, are responsible for re-paying the
+                  </ListItem>
+                  <ListItem>
+                    You, and not NFT Pawn, are responsible for re-paying the
                     amounts loaned under a Transaction;
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     Subject to Section 5(b), you acknowledge that your
                     submission of a Listing offers your NFTs, fungible tokens,
                     or other crypto-assets as collateral in exchange for a loan;
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     You will comply with all Applicable Laws in submitting each
                     loan proposal and disclosure of Borrower information in
                     connection with each loan; and
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     You have all the necessary licenses, permits, consents, or
                     approvals from or by, and have made all necessary notices
                     to, all governmental agencies having jurisdiction as
                     required by Applicable Laws.
-                  </li>
-                </ul>
-                <li>
+                  </ListItem>
+                </UnorderedList>
+                <ListItem>
                   (b) You will not be responsible for fulfilling the terms of
                   your Listing if:
-                  <ul>
-                    <li>
+                  <UnorderedList>
+                    <ListItem>
                       You cancel the Listing before it is accepted and in
-                      accordance with MyConstant's then-current cancellation
+                      accordance with NFT Pawn's then-current cancellation
                       policies; or
-                    </li>
-                    <li>Your Listing is never accepted.</li>
-                  </ul>
-                </li>
-              </li>
-            </ul>
-          </li>
-          <li>
-            <b>Lender Obligations</b>
-            <ul>
-              <li>
+                    </ListItem>
+                    <ListItem>Your Listing is never accepted.</ListItem>
+                  </UnorderedList>
+                </ListItem>
+              </ListItem>
+            </UnorderedList>
+          </ListItem>
+          <ListItem>
+            <Heading as='h2' fontSize='lg' mb={2}>Lender Obligations</Heading>
+            <UnorderedList>
+              <ListItem>
                 In addition to your other obligations under this Agreement, this
                 Section 6 is applicable to you in your capacity as a Lender.
-              </li>
-              <li>
-                You agree to comply with MyConstant's Lender policies and
+              </ListItem>
+              <ListItem>
+                You agree to comply with NFT Pawn's Lender policies and
                 practices, as set forth herein, as may be amended from time to
                 time, and agree that:
-                <ul>
-                  <li>
+                <UnorderedList>
+                  <ListItem>
                     You are responsible for familiarizing yourself with a
                     Listing and its associated loan terms;
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     You are responsible for complying with all loan terms and
                     payments of a Transaction;
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     Your acceptance of a Listing is a commitment by you to
                     provide the funds and abide by the terms of the loan from
                     the applicable Borrower;
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     You are entering into a Transaction in good faith and, if a
                     loan proposal is accepted by you, will proceed according to
                     the loan terms;
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     You will have reviewed the loan(s) for conformity with
                     Applicable Law and applicable guidelines within fifteen (15)
                     business days before your acceptance of the loan proposal,
                     or before the closing date of the loan, whichever is later;
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     You will, in a timely manner, communicate transfer
-                    instructions to the Borrower (or if required by MyConstant,
-                    to MyConstant); and
-                  </li>
-                  <li>
+                    instructions to the Borrower (or if required by NFT Pawn,
+                    to NFT Pawn); and
+                  </ListItem>
+                  <ListItem>
                     You will pay all applicable amounts to Borrower and
-                    MyConstant in association with the Transaction.
-                  </li>
-                </ul>
-              </li>
-            </ul>
-          </li>
-          <li>
-            <b>Fees; Penalties; Taxes; Audits.</b>
-            <ul>
-              <li>
+                    NFT Pawn in association with the Transaction.
+                  </ListItem>
+                </UnorderedList>
+              </ListItem>
+            </UnorderedList>
+          </ListItem>
+          <ListItem>
+            <Heading as='h2' fontSize='lg' mb={2}>Fees; Penalties; Taxes; Audits.</Heading>
+            <UnorderedList>
+              <ListItem>
                 The fees for the NFT Pawn Platform (collectively, the "Fees")
                 are as communicated to you by us from time to time, except that
                 if we have a separate written agreement setting forth such fees,
-                such agreement will control, and MyConstant reserves the right
+                such agreement will control, and NFT Pawn reserves the right
                 to modify, change, supplement or add to such Fees from time to
                 time. We will provide you with notice of any such modifications,
                 changes, supplements or additions through the NFT Pawn
                 Platform, via email, or through other reasonable means, and your
                 continued use of the NFT Pawn Platform indicates your
                 acceptance of such modifications, changes, supplements or
-                additions to the Fees. MyConstant reserves the right to require
+                additions to the Fees. NFT Pawn reserves the right to require
                 payment in advance.
-              </li>
-              <li>
+              </ListItem>
+              <ListItem>
                 You will be responsible for paying any Fees that arise out of a
                 Listing or Transaction in connection with the NFT Pawn
                 Platform, even if terms were finalized or payment was made
@@ -485,7 +368,7 @@ const TermsOfService = () => {
                 pursuant to a binding written agreement between Lender and
                 Borrower in effect prior to Lender's and/or Borrower's first
                 access to the NFT Pawn Platform (as certified by an officer of
-                Borrower and Lender upon MyConstant's request). If MyConstant
+                Borrower and Lender upon NFT Pawn's request). If NFT Pawn
                 believes that you acted to circumvent such policy, you may be
                 subject to a range of actions, including limits on your access
                 to the NFT Pawn Platform, buying and selling privileges and
@@ -493,19 +376,19 @@ const TermsOfService = () => {
                 account, financial penalties, the recovery of lost current and
                 future revenue and expenses for policy monitoring and
                 enforcement.
-              </li>
-              <li>
-                MyConstant is not a party to any Transaction, and does not
-                represent that MyConstant collects taxes, whether municipal,
+              </ListItem>
+              <ListItem>
+                NFT Pawn is not a party to any Transaction, and does not
+                represent that NFT Pawn collects taxes, whether municipal,
                 state, federal, or otherwise, with respect to any Transactions.
                 You and any other users that are parties to a Transaction are
                 responsible for any taxes, if any, with respect to such
                 Transaction. You must agree to our payment method when buying or
                 bidding through the NFT Pawn Platform, and pay all fees and
                 applicable taxes no later than when the Transaction consummates.
-              </li>
-              <li>
-                MyConstant reserves the right, upon reasonable prior notice to
+              </ListItem>
+              <ListItem>
+                NFT Pawn reserves the right, upon reasonable prior notice to
                 you and during normal business hours, to itself (or through its
                 designees) audit your usage of the NFT Pawn Platform (including
                 your buying and selling and other Transactions on the NFT Pawn
@@ -513,66 +396,66 @@ const TermsOfService = () => {
                 Agreement. If such audit reveals that you have improperly used
                 the NFT Pawn Platform, or failed to make payments as required
                 under this Agreement, such conduct shall be considered a
-                material breach of this Agreement and MyConstant may choose, in
+                material breach of this Agreement and NFT Pawn may choose, in
                 its sole discretion to terminate or suspend your access to the
                 NFT Pawn Platform and/or this Agreement, and/or invoice you for
-                such unauthorized use based upon MyConstant's standard fees in
+                such unauthorized use based upon NFT Pawn's standard fees in
                 effect at the time of the audit, which you shall promptly pay to
-                MyConstant. If the underpaid fees exceed five percent (5%) of
-                the fees actually paid, you shall also pay MyConstant's
+                NFT Pawn. If the underpaid fees exceed five percent (5%) of
+                the fees actually paid, you shall also pay NFT Pawn's
                 reasonable costs of conducting the audit.
-              </li>
-            </ul>
-          </li>
-          <li>
-            <b>Intellectual Property Rights.</b>
-            <ul>
-              <li>
+              </ListItem>
+            </UnorderedList>
+          </ListItem>
+          <ListItem>
+            <Heading as='h2' fontSize='lg' mb={2}>Intellectual Property Rights.</Heading>
+            <UnorderedList>
+              <ListItem>
                 Except for your User Material (as defined below), the NFT Pawn
                 Platform and all materials therein or made available in
                 connection therewith, including software, data, usage data, data
                 sets, images, text, graphics, charts, data, illustrations,
                 logos, trademarks, service marks, and User Material of other
-                users (the "MyConstant Material"), and all intellectual property
-                rights related thereto, are the exclusive property of MyConstant
+                users (the "NFT Pawn Material"), and all intellectual property
+                rights related thereto, are the exclusive property of NFT Pawn
                 and/or its licensors (including other users, as applicable).
                 Except as expressly permitted herein, nothing in this Agreement
                 shall be deemed to create a license with respect to any such
-                intellectual property rights. Use of the MyConstant Material for
+                intellectual property rights. Use of the NFT Pawn Material for
                 any purpose not expressly permitted by this Agreement is
                 prohibited.
-              </li>
-              <li>
+              </ListItem>
+              <ListItem>
                 Subject to the terms and conditions of this Agreement, you are
                 hereby granted a limited, non-exclusive, non-sublicensable,
                 non-transferable, revocable license to use the NFT Pawn
                 Platform during the term of this Agreement, for your internal
                 use only, and only to the extent permitted by the features and
                 functionalities of the NFT Pawn Platform.
-              </li>
-              <li>
+              </ListItem>
+              <ListItem>
                 You may choose to, or we may invite you to, submit comments or
-                ideas relating to MyConstant or the NFT Pawn Platform,
+                ideas relating to NFT Pawn or the NFT Pawn Platform,
                 including about how to improve the NFT Pawn Platform or our
                 other products and services (" Feedback"). By making available
                 any Feedback, you hereby agree that your disclosure is
                 gratuitous, unsolicited and without restriction and will not
-                place MyConstant under any agency, fiduciary or other
+                place NFT Pawn under any agency, fiduciary or other
                 obligation, and that we are free to use and otherwise exploit
                 the Feedback (and authorize others to do so) in any way
                 whatsoever without any additional compensation to you, and/or to
                 publicly disclose the Feedback to others. Without limiting the
                 foregoing, you acknowledge that, by receiving your Feedback,
-                MyConstant does not waive any rights to use similar or related
-                ideas previously known to MyConstant, or developed by its
+                NFT Pawn does not waive any rights to use similar or related
+                ideas previously known to NFT Pawn, or developed by its
                 personnel, or obtained from other sources.
-              </li>
-            </ul>
-          </li>
-          <li>
-            <b>User Material.</b>
-            <ul>
-              <li>
+              </ListItem>
+            </UnorderedList>
+          </ListItem>
+          <ListItem>
+            <Heading as='h2' fontSize='lg' mb={2}>User Material.</Heading>
+            <UnorderedList>
+              <ListItem>
                 The NFT Pawn Platform may allow users to submit, post, display,
                 provide, or otherwise make available content such as profile
                 information, information about NFTs, fungible tokens or other
@@ -580,8 +463,8 @@ const TermsOfService = () => {
                 material, messages, comments, questions, answers and other
                 content or information, including Listings (collectively, "User
                 Material").
-              </li>
-              <li>
+              </ListItem>
+              <ListItem>
                 WE CLAIM NO OWNERSHIP RIGHTS OVER USER MATERIAL CREATED BY YOU.
                 THE USER MATERIAL YOU CREATE REMAINS YOURS, SUBJECT TO THE
                 RIGHTS GRANTED IN THIS AGREEMENT. However, you understand that
@@ -589,11 +472,11 @@ const TermsOfService = () => {
                 NFT Pawn Platform, you agree to allow others to view, edit,
                 share, and/or interact with your User Material, in accordance
                 with your settings on the NFT Pawn Platform and this Agreement.
-                MyConstant has no control over and is not responsible for any
+                NFT Pawn has no control over and is not responsible for any
                 use or misuse (including any distribution) by any third party of
                 User Material.
-              </li>
-              <li>
+              </ListItem>
+              <ListItem>
                 You hereby grant to us a worldwide, royalty-free, fully paid-up,
                 non-exclusive, perpetual, irrevocable, transferable and fully
                 sublicensable (through multiple tiers) license, without
@@ -604,38 +487,38 @@ const TermsOfService = () => {
                 exploit (collectively, "Use") any of your User Material, in any
                 format or media now known or hereafter developed, and for any
                 purpose (including promotional purposes, such as testimonials).
-              </li>
-              <li>
+              </ListItem>
+              <ListItem>
                 In connection with your User Material, you represent and warrant
                 that:
-                <ul>
-                  <li>
+                <UnorderedList>
+                  <ListItem>
                     You have the written consent of each and every identifiable
                     natural person in the User Material, if any, to use such
                     person's name or likeness in the manner contemplated by the
                     NFT Pawn Platform and this Agreement, and each such person
-                    has released you and MyConstant from any liability that may
+                    has released you and NFT Pawn from any liability that may
                     arise in relation to such use.
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     You have obtained and are solely responsible for obtaining
                     all consents as may be required by Applicable Law to post
                     any User Material relating to third parties.
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     Your User Material and our use thereof as contemplated by
                     the NFT Pawn Platform and this Agreement will not violate
                     any Applicable Law or infringe any rights, including
                     intellectual property rights, of any third party.
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     All of your User Material and other information that you
                     provide to us is truthful and accurate.
-                  </li>
-                </ul>
-              </li>
-              <li>
-                MyConstant takes no responsibility and assumes no liability for
+                  </ListItem>
+                </UnorderedList>
+              </ListItem>
+              <ListItem>
+                NFT Pawn takes no responsibility and assumes no liability for
                 any User Material that you or any other user or third party
                 makes available. You shall be solely responsible for your User
                 Material and the consequences of making it available, and you
@@ -643,16 +526,16 @@ const TermsOfService = () => {
                 online distribution and publication of User Material. You agree
                 that you may be exposed to User Material that is inaccurate,
                 incomplete or otherwise unsuited to your purpose, and you agree
-                that MyConstant shall not be responsible or liable for any
+                that NFT Pawn shall not be responsible or liable for any
                 damages you allege to incur in connection with User Material. We
                 may (but have no obligation to) monitor, evaluate, alter, or
                 remove User Material before or after it appears on the NFT Pawn
                 Platform, or analyze your access to or use of the NFT Pawn
                 Platform, for any or no reason.
-              </li>
-              <li>
+              </ListItem>
+              <ListItem>
                 Nothing in this Agreement shall be deemed to be a representation
-                or warranty by MyConstant with respect to any Third Party
+                or warranty by NFT Pawn with respect to any Third Party
                 Materials. We have no obligation to monitor Third Party
                 Materials, and we may block or disable access to any Third Party
                 Materials (in whole or part) through the NFT Pawn Platform at
@@ -665,25 +548,25 @@ const TermsOfService = () => {
                 Platform, or share your User Material on or through any
                 third-party website or service, you do so at your own risk, and
                 you understand that this Agreement does not apply to your use of
-                such Third Party Materials. You expressly relieve MyConstant
+                such Third Party Materials. You expressly relieve NFT Pawn
                 from any and all liability arising from your use of any Third
                 Party Materials.
-              </li>
-              <li>
+              </ListItem>
+              <ListItem>
                 YOUR USE OF THIRD PARTY MATERIALS IS AT YOUR OWN RISK AND IS
                 SUBJECT TO ANY ADDITIONAL TERMS, CONDITIONS AND POLICIES
                 APPLICABLE TO SUCH THIRD PARTY MATERIALS (SUCH AS TERMS OF
                 SERVICE OR PRIVACY POLICIES OF THE PROVIDERS OF SUCH THIRD PARTY
                 MATERIALS).
-              </li>
-            </ul>
-          </li>
-          <li>
-            <b>Compliance with Laws.</b>
-            <ul>
-              <li>
+              </ListItem>
+            </UnorderedList>
+          </ListItem>
+          <ListItem>
+            <Heading as='h2' fontSize='lg' mb={2}>Compliance with Laws.</Heading>
+            <UnorderedList>
+              <ListItem>
                 The NFT Pawn Platform is controlled or operated (or both) from
-                the United States, and is not intended to subject MyConstant to
+                the United States, and is not intended to subject NFT Pawn to
                 any non-U.S. jurisdiction or law. The NFT Pawn Platform may not
                 be appropriate or available for use in some non-U.S.
                 jurisdictions. Any use of the NFT Pawn Platform is at your own
@@ -691,22 +574,22 @@ const TermsOfService = () => {
                 We may limit the NFT Pawn Platform's availability at any time,
                 in whole or in part, to any person, geographic area or
                 jurisdiction that we choose.
-              </li>
-              <li>
-                In no event shall MyConstant be obligated to take any action
-                would cause MyConstant to be in violation of any Applicable Laws
-                or would result in MyConstant being an escrow agent, debt
+              </ListItem>
+              <ListItem>
+                In no event shall NFT Pawn be obligated to take any action
+                would cause NFT Pawn to be in violation of any Applicable Laws
+                or would result in NFT Pawn being an escrow agent, debt
                 collector or consumer reporting agency under the Fair Credit
                 Reporting Act. If a change in Applicable Laws makes the
                 continued operation of any part of the NFT Pawn Platform, in
                 our sole discretion, unduly burdensome or unlawful or if a
-                regulatory authority directs MyConstant to suspend or cease
-                operating the NFT Pawn Platform, MyConstant may cease offering
+                regulatory authority directs NFT Pawn to suspend or cease
+                operating the NFT Pawn Platform, NFT Pawn may cease offering
                 the NFT Pawn Platform and terminate this Agreement upon notice
                 to you, and without further obligation to you, after which
-                MyConstant will return to you any applicable prepaid fees.
-              </li>
-              <li>
+                NFT Pawn will return to you any applicable prepaid fees.
+              </ListItem>
+              <ListItem>
                 You are responsible for complying with United States export
                 controls and for any violation of such controls, including any
                 United States embargoes or other federal rules and regulations
@@ -715,44 +598,44 @@ const TermsOfService = () => {
                 country subject to a U.S. government embargo or other
                 restriction, or that has been designated by the U.S. government
                 as a "terrorist supporting" country; or (b) on any of the U.S.
-                government lists of restricted end users. MyConstant will not be
+                government lists of restricted end users. NFT Pawn will not be
                 required, either directly or indirectly, to provide goods,
                 software, services and/or technical data that may be prohibited
                 by export control or sanctions Applicable Laws.
-              </li>
-            </ul>
-          </li>
-          <li>
-            <b>Warranties</b>
-            <ul>
-              <li>
+              </ListItem>
+            </UnorderedList>
+          </ListItem>
+          <ListItem>
+            <Heading as='h2' fontSize='lg' mb={2}>Warranties</Heading>
+            <UnorderedList>
+              <ListItem>
                 You represent and warrant that:
-                <ul>
-                  <li>
+                <UnorderedList>
+                  <ListItem>
                     you have obtained all the necessary licenses, permits,
                     consents, or approvals from or by, and have made all
                     necessary notices to, all government agencies (whether
                     federal, state, municipal or local) as necessary to fully
                     perform under this Agreement, submit Listings, consummate
-                    Transactions, Use, and allow MyConstant to Use, your User
+                    Transactions, Use, and allow NFT Pawn to Use, your User
                     Material, and as required by Applicable Law;
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     you are authorized by your Organization to conduct business
                     on the Organization's behalf and to bind such Organization
                     to the terms of this Agreement;
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     you have the right to consent to this Agreement and act in
                     accordance with its terms; the acceptance of and performance
                     of this Agreement are within your corporate powers;
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     you have been duly authorized by all necessary corporate
                     action and do not require any further consent of or filing
                     with any third party or governmental body or agency;
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     you are (a) duly organized, validly existing and in good
                     standing under the laws of the jurisdiction of your
                     organization; (b) qualified to transact business; (c)
@@ -767,43 +650,43 @@ const TermsOfService = () => {
                     jurisdiction to the extent necessary to ensure the
                     enforceability of each mortgage note and the sale of the
                     mortgage loans as contemplated by this Agreement;
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     you are able to, and will, perform each and every covenant
                     in this Agreement;
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     you are solvent and the acceptance of a Listing will not
                     cause you to become insolvent;
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     the submission of a Listing or consummation of a Transaction
                     is not undertaken with the intent to hinder, delay or
                     defraud anyone, is made in good faith, and is not fraudulent
                     or intended to deceive;
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     you have not filed a petition in bankruptcy or for
                     protection under any other federal or state debtor relief
-                    law during your time of engagement with MyConstant, and you
-                    will immediately inform MyConstant in writing if this status
+                    law during your time of engagement with NFT Pawn, and you
+                    will immediately inform NFT Pawn in writing if this status
                     changes; and
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     there is no action, suit, proceeding, investigation,
                     litigation or threat thereof that, either in any one
                     instance or in the aggregate, would adversely affect any
                     Transaction, or your ability to perform under this Agreement
-                  </li>
-                </ul>
-              </li>
-            </ul>
-          </li>
-          <li>
-            <b>Indemnity</b>
-            <ul>
-              <li>
-                You agree to defend, indemnify and hold harmless MyConstant and
+                  </ListItem>
+                </UnorderedList>
+              </ListItem>
+            </UnorderedList>
+          </ListItem>
+          <ListItem>
+            <Heading as='h2' fontSize='lg' mb={2}>Indemnity</Heading>
+            <UnorderedList>
+              <ListItem>
+                You agree to defend, indemnify and hold harmless NFT Pawn and
                 the Affiliated Entities, and their respective successors and
                 assigns, from and against any and all damages, obligations,
                 losses, liabilities, costs, debt, and expenses (including
@@ -823,30 +706,30 @@ const TermsOfService = () => {
                 other user of the NFT Pawn Platform; and/or (h) any other
                 party's access to or use of the NFT Pawn Platform with your
                 username, password or other appropriate security code
-              </li>
-            </ul>
-          </li>
-          <li>
-            <b>DISCLAIMER</b>
-            <ul>
-              <li>
+              </ListItem>
+            </UnorderedList>
+          </ListItem>
+          <ListItem>
+            <Heading as='h2' fontSize='lg' mb={2}>DISCLAIMER</Heading>
+            <UnorderedList>
+              <ListItem>
                 TO THE FULLEST EXTENT PERMITTED UNDER APPLICABLE LAW: (A) THE
                 PLATFORM AND ANY THIRD PARTY MATERIALS (INCLUDING ALL LISTINGS
                 AND OTHER USER MATERIAL) ARE MADE AVAILABLE ON AN "AS IS,"
                 "WHERE IS" AND "WHERE AVAILABLE" BASIS, WITHOUT ANY WARRANTIES
                 OF ANY KIND, WHETHER EXPRESS, IMPLIED OR STATUTORY; AND (B)
-                MyConstant DISCLAIMS ALL WARRANTIES WITH RESPECT TO THE PLATFORM
+                NFT Pawn DISCLAIMS ALL WARRANTIES WITH RESPECT TO THE PLATFORM
                 AND ANY TRANSACTIONS OR THIRD PARTY MATERIALS, INCLUDING THE
                 WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE,
                 NON-INFRINGEMENT AND TITLE. ALL DISCLAIMERS OF ANY KIND
                 (INCLUDING IN THIS SECTION AND ELSEWHERE IN THIS AGREEMENT) ARE
-                MADE FOR THE BENEFIT OF BOTH MyConstant AND ITS AFFILIATES, AND
+                MADE FOR THE BENEFIT OF BOTH NFT Pawn AND ITS AFFILIATES, AND
                 ITS AND THEIR RESPECTIVE SHAREHOLDERS, DIRECTORS, OFFICERS,
                 EMPLOYEES, AFFILIATES, AGENTS, REPRESENTATIVES, LICENSORS,
                 SUPPLIERS AND SERVICE PROVIDERS (COLLECTIVELY, THE "AFFILIATED
                 ENTITIES"), AND THEIR RESPECTIVE SUCCESSORS AND ASSIGNS.
-              </li>
-              <li>
+              </ListItem>
+              <ListItem>
                 While we try to maintain the timeliness, integrity and security
                 of the NFT Pawn Platform, we do not guarantee that the
                 NFT Pawn Platform is or will remain updated, complete, correct
@@ -855,17 +738,17 @@ const TermsOfService = () => {
                 errors and materials that violate or conflict with this
                 Agreement. Additionally, third parties may make unauthorized
                 alterations to the NFT Pawn Platform. If you become aware of
-                any such alteration, contact us at hello@myconstant.com with a
+                any such alteration, contact us at hello@nftpawn.financial with a
                 description of such alteration and its location.
-              </li>
-            </ul>
-          </li>
-          <li>
+              </ListItem>
+            </UnorderedList>
+          </ListItem>
+          <ListItem>
             LIMITATION OF LIABILITY.
-            <ul>
-              <li>
+            <UnorderedList>
+              <ListItem>
                 TO THE FULLEST EXTENT PERMITTED UNDER APPLICABLE LAW: (A)
-                MyConstant WILL NOT BE LIABLE FOR ANY INDIRECT, INCIDENTAL,
+                NFT Pawn WILL NOT BE LIABLE FOR ANY INDIRECT, INCIDENTAL,
                 CONSEQUENTIAL, SPECIAL, EXEMPLARY OR PUNITIVE DAMAGES OF ANY
                 KIND, UNDER ANY CONTRACT, TORT (INCLUDING NEGLIGENCE), STRICT
                 LIABILITY OR OTHER THEORY, INCLUDING DAMAGES FOR LOSS OF
@@ -873,34 +756,34 @@ const TermsOfService = () => {
                 SECURITY OF USER MATERIAL (INCLUDING UNAUTHORIZED INTERCEPTION
                 BY THIRD PARTIES OF ANY USER MATERIAL), EVEN IF ADVISED IN
                 ADVANCE OF THE POSSIBILITY OF SUCH DAMAGES OR LOSSES; (B)
-                WITHOUT LIMITING THE FOREGOING, MyConstant WILL NOT BE LIABLE
+                WITHOUT LIMITING THE FOREGOING, NFT Pawn WILL NOT BE LIABLE
                 FOR DAMAGES OF ANY KIND RESULTING FROM YOUR USE OF OR INABILITY
                 TO USE THE PLATFORM OR FROM ANY THIRD PARTY MATERIALS, INCLUDING
                 FROM ANY VIRUS THAT MAY BE TRANSMITTED IN CONNECTION THEREWITH;
                 (C) YOUR SOLE AND EXCLUSIVE REMEDY FOR DISSATISFACTION WITH THE
                 PLATFORM OR THIRD PARTY MATERIALS IS TO STOP USING THE PLATFORM;
-                AND (D) THE MAXIMUM AGGREGATE LIABILITY OF MyConstant FOR ALL
+                AND (D) THE MAXIMUM AGGREGATE LIABILITY OF NFT Pawn FOR ALL
                 DAMAGES, LOSSES AND CAUSES OF ACTION, WHETHER IN CONTRACT, TORT
                 (INCLUDING NEGLIGENCE) OR OTHERWISE, SHALL BE THE LESSER OF
-                ($100) OR THE FEES PAID TO MyConstant IN CONNECTION WITH THE
+                ($100) OR THE FEES PAID TO NFT Pawn IN CONNECTION WITH THE
                 TRANSACTION(S) GIVING RISE TO THE LIABILITY. ALL LIMITATIONS OF
                 LIABILITY OF ANY KIND (INCLUDING IN THIS SECTION AND ELSEWHERE
-                IN THIS AGREEMENT) ARE MADE FOR THE BENEFIT OF BOTH MyConstant
+                IN THIS AGREEMENT) ARE MADE FOR THE BENEFIT OF BOTH NFT Pawn
                 AND THE AFFILIATED ENTITIES, AND THEIR RESPECTIVE SUCCESSORS AND
                 ASSIGNS.
-              </li>
-              <li>
+              </ListItem>
+              <ListItem>
                 No action, regardless of form, arising out of this Agreement may
                 be brought by you more than one (1) year after you knew or
                 should have known of the event which gave rise to the cause of
                 action.
-              </li>
-            </ul>
-          </li>
-          <li>
-            <b>Termination;Suspension.</b>
-            <ul>
-              <li>
+              </ListItem>
+            </UnorderedList>
+          </ListItem>
+          <ListItem>
+            <Heading as='h2' fontSize='lg' mb={2}>Termination;Suspension.</Heading>
+            <UnorderedList>
+              <ListItem>
                 The term of this Agreement will continue until terminated in
                 accordance with its terms. The following provisions shall
                 survive termination of this Agreement: Section 4 through Section
@@ -908,8 +791,8 @@ const TermsOfService = () => {
                 such termination), Section 7 (with respect to any Fees accrued
                 before such termination or during such one (1)-year period set
                 forth in Section 7), and Section 8 through Section 23.
-              </li>
-              <li>
+              </ListItem>
+              <ListItem>
                 In addition to our other termination rights under this
                 Agreement, we may terminate this Agreement immediately upon
                 written notice to you if you commit any material breach of this
@@ -918,25 +801,25 @@ const TermsOfService = () => {
                 and the proceedings are not dismissed within sixty (60) days, or
                 otherwise become generally unable to meet your obligations under
                 this Agreement.
-              </li>
-              <li>
+              </ListItem>
+              <ListItem>
                 Upon any such termination:
-                <ul>
-                  <li>
+                <UnorderedList>
+                  <ListItem>
                     Your right to access and use the NFT Pawn Platform will
                     cease immediately;
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     You will pay any fees accrued prior to such termination, and
                     for partial months or partial performance, any applicable
                     fees will be pro-rated or otherwise accounted for;
-                  </li>
-                  <li>
-                    MyConstant may (but is not obligated to) provide you with
+                  </ListItem>
+                  <ListItem>
+                    NFT Pawn may (but is not obligated to) provide you with
                     access to, or the ability to export, all or a portion of
                     your User Material; and
-                  </li>
-                  <li>
+                  </ListItem>
+                  <ListItem>
                     Any Confidential Information (as defined below) provided in
                     connection with the Agreement shall be promptly destroyed by
                     the receiving party (or, at the disclosing party's written
@@ -946,39 +829,39 @@ const TermsOfService = () => {
                     such party's retention policies, provided that the receiving
                     party shall continue to treat it as Confidential
                     Information.
-                  </li>
-                  <li>
-                    MyConstant may terminate or suspend your right to access or
+                  </ListItem>
+                  <ListItem>
+                    NFT Pawn may terminate or suspend your right to access or
                     use all or any portion of the NFT Pawn Platform at any
                     time, and for any or no reason, including if:
-                    <ul>
-                      <li>
-                        MyConstant believes that you (a) pose a security risk
+                    <UnorderedList>
+                      <ListItem>
+                        NFT Pawn believes that you (a) pose a security risk
                         to, or may materially harm,
-                      </li>
-                      <li>
-                        MyConstant, the NFT Pawn Platform, or any user or other
+                      </ListItem>
+                      <ListItem>
+                        NFT Pawn, the NFT Pawn Platform, or any user or other
                         third party; (b) may adversely impact the services,
                         systems or content of any other user; or (c) may subject
-                        MyConstant, or any user or other third party to
-                        liability; MyConstant determines or reasonably suspects
+                        NFT Pawn, or any user or other third party to
+                        liability; NFT Pawn determines or reasonably suspects
                         that you may have violated this Agreement; or
-                      </li>
-                      <li>
-                        MyConstant has been notified by a third party that you
+                      </ListItem>
+                      <ListItem>
+                        NFT Pawn has been notified by a third party that you
                         have violated such third party's terms and conditions in
                         connection with the NFT Pawn Platform.
-                      </li>
-                    </ul>
-                  </li>
-                </ul>
-              </li>
-            </ul>
-          </li>
-          <li>
-            <b>Confidentiality</b>
-            <ul>
-              <li>
+                      </ListItem>
+                    </UnorderedList>
+                  </ListItem>
+                </UnorderedList>
+              </ListItem>
+            </UnorderedList>
+          </ListItem>
+          <ListItem>
+            <Heading as='h2' fontSize='lg' mb={2}>Confidentiality</Heading>
+            <UnorderedList>
+              <ListItem>
                 Each of you and us may have access to information (in any form)
                 that relates to the others' past, present, and future activities
                 including research, development, business activities, products,
@@ -989,7 +872,7 @@ const TermsOfService = () => {
                 receiving party consistent with the rights and obligations of
                 this Agreement. Our Confidential Information includes a
                 software, NFT Pawn Platform, and any related documentation made
-                available by MyConstant to you. The receiving party agrees to
+                available by NFT Pawn to you. The receiving party agrees to
                 protect the Confidential Information of the disclosing party in
                 the same manner that it protects its own similar confidential
                 information, but in no event using less than a reasonable
@@ -1013,13 +896,13 @@ const TermsOfService = () => {
                 disclose for any purpose, on an anonymized or aggregated basis,
                 any data collected in connection with use of the NFT-Lend
                 Platform, including any Listings or Transactions.
-              </li>
-            </ul>
-          </li>
-          <li>
-            <b>Publicity</b>
-            <ul>
-              <li>
+              </ListItem>
+            </UnorderedList>
+          </ListItem>
+          <ListItem>
+            <Heading as='h2' fontSize='lg' mb={2}>Publicity</Heading>
+            <UnorderedList>
+              <ListItem>
                 Neither party may use the names, trade names, trademarks, logos,
                 service marks or trade dress of the other party in any manner
                 other than as expressly set forth herein, including in
@@ -1029,34 +912,34 @@ const TermsOfService = () => {
                 the NFT Pawn Platform, without such party's prior written
                 consent, which may be withheld or granted in its sole
                 discretion. Notwithstanding the foregoing, you grant to
-                MyConstant the limited, non-exclusive right to use your name,
+                NFT Pawn the limited, non-exclusive right to use your name,
                 logos, and trademarks (collectively, "User Marks"): (a) on
-                internal and public materials (including MyConstant's website)
+                internal and public materials (including NFT Pawn's website)
                 to identify you as a user of the NFT Pawn Platform; and (b) as
                 part of authorized press releases (as described below). You
-                agree that MyConstant may issue a press release after the
+                agree that NFT Pawn may issue a press release after the
                 effective date of this Agreement that announces the business
-                relationship formed by this Agreement. MyConstant agrees that
+                relationship formed by this Agreement. NFT Pawn agrees that
                 its use of User Marks, including the goodwill and reputation
                 associated therewith, shall be in accordance with your generally
                 applicable trademark guidelines. Unless otherwise agreed in
-                writing, MyConstant will cease using the User Marks as provided
+                writing, NFT Pawn will cease using the User Marks as provided
                 in this Section within a commercially reasonable period after
                 termination of this Agreement.
-              </li>
-            </ul>
-          </li>
-          <li>
-            <b>Governing Law; Arbitration.</b>
-            <ul>
-              <li>
+              </ListItem>
+            </UnorderedList>
+          </ListItem>
+          <ListItem>
+            <Heading as='h2' fontSize='lg' mb={2}>Governing Law; Arbitration.</Heading>
+            <UnorderedList>
+              <ListItem>
                 (a) This Agreement and all rights and obligations hereunder,
                 including matters of construction, validity and performance, and
                 your use of the NFT Pawn Platform, shall be governed by and
                 construed in accordance with the laws of the State of New York
                 without regard to its conflict of laws provisions.
-              </li>
-              <li>
+              </ListItem>
+              <ListItem>
                 (b) Except as herein specifically stated, any dispute, claim or
                 controversy arising out of or relating to statutory or common
                 law claims, the breach, termination, enforcement, interpretation
@@ -1078,15 +961,15 @@ const TermsOfService = () => {
                 arbitration and the terms of the Arbitrable Dispute shall be
                 kept confidential by the parties, except as may be required by
                 Applicable Law.
-              </li>
-              <li>
+              </ListItem>
+              <ListItem>
                 (c) The arbitrator shall be mutually agreed upon by the parties.
                 If the parties are unable to agree within twenty (20) days
                 following submission of the dispute to JAMS by one of the
                 parties, JAMS will have the authority to select an arbitrator in
                 accordance with the JAMS Rules.
-              </li>
-              <li>
+              </ListItem>
+              <ListItem>
                 (d) No discovery other than an exchange of relevant documents
                 may occur in any arbitration commenced under the provisions of
                 this Agreement. The parties agree to act in good faith to
@@ -1096,8 +979,8 @@ const TermsOfService = () => {
                 designation and to render an award no later than ten (10) days
                 after the conclusion of such hearing, in each case unless
                 otherwise mutually agreed in writing by the parties.
-              </li>
-              <li>
+              </ListItem>
+              <ListItem>
                 (e) The final decision of arbitrator shall (i) identify the
                 prevailing party and state the amount of the award (the "Award
                 Amount"), if any; (ii) include a reasonably detailed written
@@ -1108,65 +991,64 @@ const TermsOfService = () => {
                 confidential unless otherwise agreed in writing by the parties,
                 and an order with respect thereto may be entered in any court of
                 competent jurisdiction.
-              </li>
-              <li>
+              </ListItem>
+              <ListItem>
                 (f) The arbitrator will not have the power to alter, amend or
                 otherwise affect the terms of these arbitration provisions or
                 any other provisions contained in this Section or elsewhere in
                 this Agreement.
-              </li>
-            </ul>
-          </li>
-          <li>
-            <b>Assignment.</b>
-            <ul>
-              <li>
+              </ListItem>
+            </UnorderedList>
+          </ListItem>
+          <ListItem>
+            <Heading as='h2' fontSize='lg' mb={2}>Assignment.</Heading>
+            <UnorderedList>
+              <ListItem>
                 You may not assign this Agreement or delegate, in each case, any
                 of your rights or obligations hereunder without the prior
-                written consent of MyConstant. Any attempted assignment in
-                violation of this Section 20 shall be void. MyConstant may
+                written consent of NFT Pawn. Any attempted assignment in
+                violation of this Section 20 shall be void. NFT Pawn may
                 assign this Agreement or delegate any of its rights or
                 obligations hereunder without your consent. Subject to the
                 foregoing limitations, this Agreement will be binding upon and
                 inure to the benefit of the parties and their respective
                 successors and permitted assigns.
-              </li>
-            </ul>
-          </li>
-          <li>
-            <b>Copyright Infringement Claims.</b>
-            <ul>
-              <li>
+              </ListItem>
+            </UnorderedList>
+          </ListItem>
+          <ListItem>
+            <Heading as='h2' fontSize='lg' mb={2}>Copyright Infringement Claims.</Heading>
+            <UnorderedList>
+              <ListItem>
                 The Digital Millennium Copyright Act of 1998 (the "DMCA")
                 provides recourse for copyright owners who believe that material
                 appearing on the Internet infringes their rights under U.S.
                 copyright law. If you believe in good faith that material
                 available on the NFT Pawn Platform infringes your copyright,
-                you (or your agent) may send to MyConstant a written notice by
+                you (or your agent) may send to NFT Pawn a written notice by
                 mail or e-mail, requesting that we remove such material or block
                 access to it. If you believe in good faith that someone has
                 wrongly filed a notice of copyright infringement against you,
-                the DMCA permits you to send to MyConstant a counter-notice.
+                the DMCA permits you to send to NFT Pawn a counter-notice.
                 Notices and counter-notices must meet the then-current statutory
                 requirements imposed by the DMCA. See{" "}
                 <a target={"_blank"} href="http://www.copyright.gov/">
                   http://www.copyright.gov/
                 </a>
                 for details. Notices and counter-notices must be sent in writing
-                to Robert Masiello as follows: By mail to 1390 Market Street
-                Suite 200 San Francisco, CA 94102; or by e-mail to
-                <a href="mailto:hello@myconstant.com">hello@myconstant.com</a>.
-              </li>
-              <li>
+                to Zon Chu as follows: By mail to 3500 S dupont hwy Dover 19901 Delaware; or by e-mail to
+                <a href="mailto:hello@nftpawn.financial">hello@nftpawn.financial</a>.
+              </ListItem>
+              <ListItem>
                 We suggest that you consult your legal advisor before filing a
                 DMCA notice or counter-notice.
-              </li>
-            </ul>
-          </li>
-          <li>
-            <b>Notices; Electronic Consent.</b>
-            <ul>
-              <li>
+              </ListItem>
+            </UnorderedList>
+          </ListItem>
+          <ListItem>
+            <Heading as='h2' fontSize='lg' mb={2}>Notices; Electronic Consent.</Heading>
+            <UnorderedList>
+              <ListItem>
                 You consent to electronic provision of all notices,
                 communications and disclosures ("Communications") from us,
                 including those required by Applicable Laws. You agree that your
@@ -1177,44 +1059,44 @@ const TermsOfService = () => {
                 considered received 24 hours after we deliver them to you. You
                 may withdraw your consent to receive electronic Communications
                 at any time by sending us a written request to
-                hello@myconstant.com. Due to the nature of the NFT Pawn
+                hello@nftpawn.financial. Due to the nature of the NFT Pawn
                 Platform, you will not be able to continue to use and/or access
                 the NFT Pawn Platform without agreeing to electronic delivery
                 of Communications, and if you elect to withdraw your consent to
                 receive Communications electronically you will no longer have
                 access to or be able to use the NFT Pawn Platform.
-              </li>
-              <li>
+              </ListItem>
+              <ListItem>
                 All notices, demands and other communications you provide to us
                 must be in writing and must be sent by electronic mail to the
-                following address: hello@myconstant.com (or to such other
-                address as hello@myconstant.com may designate to you from time
+                following address: hello@nftpawn.financial (or to such other
+                address as hello@nftpawn.financial may designate to you from time
                 to time).
-              </li>
-            </ul>
-          </li>
-          <li>
-            <b>Miscellaneous</b>
-            <ul>
-              <li>
+              </ListItem>
+            </UnorderedList>
+          </ListItem>
+          <ListItem>
+            <Heading as='h2' fontSize='lg' mb={2}>Miscellaneous</Heading>
+            <UnorderedList>
+              <ListItem>
                 Nothing in this Agreement shall be deemed to create a joint
                 venture, partnership, or agency relationship between the
                 parties, or be deemed to authorize either party to incur any
                 liabilities or obligations on behalf of, or in the name of, the
                 other party.
-              </li>
-              <li>
+              </ListItem>
+              <ListItem>
                 This Agreement, including any terms and conditions incorporated
-                herein, is the entire agreement between you and MyConstant
+                herein, is the entire agreement between you and NFT Pawn
                 relating to the subject matter hereof, and supersedes any and
                 all prior or contemporaneous written or oral agreements or
-                understandings between you and MyConstant relating to such
+                understandings between you and NFT Pawn relating to such
                 subject matter.
-              </li>
-              <li>
-                MyConstant shall not be liable for any delay or failure to
+              </ListItem>
+              <ListItem>
+                NFT Pawn shall not be liable for any delay or failure to
                 perform its obligations hereunder resulting from any cause
-                beyond MyConstant's reasonable control that makes performance
+                beyond NFT Pawn's reasonable control that makes performance
                 impossible, including: fires; explosions; pandemics, epidemics
                 or other spread of disease; floods; strikes; work stoppages or
                 slowdowns or other industrial disputes; accidents; riots or
@@ -1222,19 +1104,19 @@ const TermsOfService = () => {
                 industry-wide material shortages, power failures, or
                 communication line interruptions; disruption or damage to, or
                 failure of, the Internet or other computer networks, third party
-                information technology systems or services beyond MyConstant's
-                reasonable control. You acknowledge that MyConstant does not and
+                information technology systems or services beyond NFT Pawn's
+                reasonable control. You acknowledge that NFT Pawn does not and
                 cannot control the flow of data through the Internet.
-              </li>
-              <li>
+              </ListItem>
+              <ListItem>
                 If the software in connection with the NFT Pawn Platform is
                 construed as "commercial computer software", as that term is
                 defined in 48 CFR 2.101, all U.S. government end users acquiring
                 access to such software may only do so with the rights set forth
                 in this Agreement, in accordance with 48 CFR 12.212(b) and/ or
                 48 CFR 227.7202-1(a) and 48 CFR 227.7202-4, as applicable.
-              </li>
-              <li>
+              </ListItem>
+              <ListItem>
                 We hereby notify you that parental control protections (such as
                 computer hardware, software, or filtering services) are
                 commercially available that may assist you in limiting access to
@@ -1246,8 +1128,8 @@ const TermsOfService = () => {
                 ></a>
                 . Please note that we do not endorse any of the products or
                 services listed on such site.
-              </li>
-              <li>
+              </ListItem>
+              <ListItem>
                 You may not, for the duration of this Agreement and for a period
                 of twelve (12) months thereafter, directly or indirectly solicit
                 for employment, nor offer employment to, nor enter into any
@@ -1257,14 +1139,14 @@ const TermsOfService = () => {
                 advertising, job fairs and similar means, or employment or
                 offers resulting from such solicitation, shall not be deemed to
                 be a violation of this provision.
-              </li>
-              <li>
+              </ListItem>
+              <ListItem>
                 No delay, neglect or forbearance on the part of either party in
                 enforcing any term of this Agreement shall be deemed to be a
                 waiver. All waivers must be in writing and must be signed by the
                 party against which it is sought to be enforced to be effective.
-              </li>
-              <li>
+              </ListItem>
+              <ListItem>
                 The invalidity or unenforceability of any provision of this
                 Agreement shall not affect the validity or enforceability of any
                 other provision hereof, and it is the intent and agreement of
@@ -1274,8 +1156,8 @@ const TermsOfService = () => {
                 such modification is not possible, by substituting another
                 provision that is legal and enforceable and that achieves the
                 same objective.
-              </li>
-              <li>
+              </ListItem>
+              <ListItem>
                 Any heading, caption or section title contained herein is for
                 convenience only, and in no way defines or explains any section
                 or provision. All terms defined in the singular shall have the
@@ -1283,21 +1165,21 @@ const TermsOfService = () => {
                 unless otherwise specified. Any use of the term "including" or
                 variations thereof in this Agreement shall be construed as if
                 followed by the phrase "without limitation."
-              </li>
-              <li>
+              </ListItem>
+              <ListItem>
                 This Agreement does not create any third party beneficiary
                 rights in any individual or entity that is not a party to this
                 Agreement, except that the Affiliated Entities are express
                 third-party beneficiaries of this Agreement.
-              </li>
-              <li>
-                NFT Pawn Platform © 2022 MyConstant unless otherwise noted. All
+              </ListItem>
+              <ListItem>
+                NFT Pawn Platform © 2022 NFT Pawn unless otherwise noted. All
                 rights reserved.
-              </li>
-            </ul>
-          </li>
-        </ol>
-      </div>
+              </ListItem>
+            </UnorderedList>
+          </ListItem>
+        </OrderedList>
+      </Flex>
     </BodyContainer>
   );
 };
